refactor(queue): extract live CI-aggregate/merge-state resolution into its own module
Part of #4013's module-split sequence (step 6, after transient-locks.ts,
signal-snapshot.ts, duplicate-detection.ts, slop-detection.ts, and
review-evasion.ts): the live CI-aggregate/merge-state caching and
resolution functions move to src/queue/ci-resolution.ts. Pure
mechanical move, no behavior change.

LiveGithubFacts (the request-scoped memo these functions read/write)
stays in processors.ts -- it is a foundational type used across dozens
of unrelated call sites there -- so this file imports it back
type-only, which creates no runtime circular dependency even though
processors.ts also imports several functions from this file for its
own remaining call sites. liveFactKey/liveFactTokenPart and
cachedRequiredStatusContexts are exported for the same reason: each
still has one remaining caller left in processors.ts.

Isolating this code surfaced a genuine, previously-invisible coverage
gap: cachedLiveCiAggregate's own request-scoped cache-hit branch (a
second call within the same webhook pass reusing the first call's
in-flight/settled promise) had no direct test. Added one, importing
the function directly rather than threading a full webhook simulation
through two call sites that happen to share a cache key.
